Created home page layout and functionality for wagers
Signed-off-by: Sameer <sam.chaturvedi24@gmail.com>

Added list of cards displaying each wager using RecyclerView and
CardView. At the moment, wagers are hardcoded, each wager having its own
instance of a Wager class. Clicking on each wager displays the
description of the wager as a Toast. Added temporary login page which is
just a button to get to the home page.
Also changed color scheme of app, and added fonts and images.